Move order and RMA stuff in its own classes and some other stuff too.
