+++ src/main/java/ro/sellfluence/db/EmagMirrorDB.java
@@ -2,24 +2,17 @@ package ro.sellfluence.db;
 
 import ch.claudio.db.DB;
 import ro.sellfluence.apphelper.EmployeeSheetData;
+import ro.sellfluence.db.EmagFetchLog.EmagFetchHistogram;
 import ro.sellfluence.db.ProductTable.ProductInfo;
 import ro.sellfluence.db.versions.SetupDB;
-import ro.sellfluence.emagapi.AWB;
-import ro.sellfluence.emagapi.Attachment;
 import ro.sellfluence.emagapi.CancellationReason;
-import ro.sellfluence.emagapi.Customer;
-import ro.sellfluence.emagapi.Flag;
 import ro.sellfluence.emagapi.LockerDetails;
 import ro.sellfluence.emagapi.OrderResult;
-import ro.sellfluence.emagapi.Product;
 import ro.sellfluence.emagapi.RMAResult;
-import ro.sellfluence.emagapi.RequestHistory;
-import ro.sellfluence.emagapi.ReturnedProduct;
-import ro.sellfluence.emagapi.StatusHistory;
-import ro.sellfluence.emagapi.StatusRequest;
-import ro.sellfluence.emagapi.Voucher;
-import ro.sellfluence.emagapi.VoucherSplit;
 import ro.sellfluence.sheetSupport.Conversions;
+import ro.sellfluence.test.dbexplorer.CustomerRecord;
+import ro.sellfluence.test.dbexplorer.OrderRecord;
+import ro.sellfluence.test.dbexplorer.ProductInOrderRecord;
 
 import java.io.IOException;
 import java.math.BigDecimal;
@@ -34,9 +27,7 @@ import java.util.Comparator;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
-import java.util.Objects;
 import java.util.Optional;
-import java.util.Scanner;
 import java.util.SortedMap;
 import java.util.TreeMap;
 import java.util.UUID;
@@ -44,22 +35,24 @@ import java.util.logging.Logger;
 import java.util.stream.Stream;
 
 import static java.math.RoundingMode.HALF_EVEN;
-import static java.sql.Statement.RETURN_GENERATED_KEYS;
-import static java.util.logging.Level.FINE;
 import static java.util.logging.Level.WARNING;
 import static ro.sellfluence.db.EmagFetchLog.deleteFetchLogsBefore;
 import static ro.sellfluence.db.EmagFetchLog.getEmagLog;
 import static ro.sellfluence.db.EmagFetchLog.insertEmagLog;
 import static ro.sellfluence.db.EmagFetchLog.updateEmagLog;
+import static ro.sellfluence.db.EmagOrder.addOrderResult;
 import static ro.sellfluence.db.GMV.addToGMV;
 import static ro.sellfluence.db.GMV.getGMVByMonth;
 import static ro.sellfluence.db.GMV.getGMVByProductId;
 import static ro.sellfluence.db.GMV.insertOrUpdateGMV;
 import static ro.sellfluence.db.GMV.specialCase;
 import static ro.sellfluence.db.GMV.subtractFromGMV;
+import static ro.sellfluence.db.ProductTable.getProductCodes;
 import static ro.sellfluence.db.ProductTable.getProducts;
 import static ro.sellfluence.db.ProductTable.insertProduct;
 import static ro.sellfluence.db.ProductTable.updateProduct;
+import static ro.sellfluence.db.RMA.addRMAResult;
+import static ro.sellfluence.db.Vendor.insertOrUpdateVendor;
 import static ro.sellfluence.db.Vendor.selectFetchTimeByAccount;
 import static ro.sellfluence.db.Vendor.selectVendorIdByName;
 import static ro.sellfluence.db.Vendor.updateFetchTimeByAccount;
@@ -106,339 +99,10 @@ public class EmagMirrorDB {
      * @throws SQLException if something goes wrong.
      */
     public void addOrder(OrderResult order, String account) throws SQLException {
-        var vendorName = order.vendor_name();
-        var vendorId = getOrAddVendor(vendorName, account); // Keep vendorId for the vendor table link
         database.writeTX(db -> {
-            if (order.customer() != null) {
-                insertOrUpdateCustomer(db, order.customer());
-            }
-            if (order.details() != null && order.details().locker_id() != null) {
-                insertOrUpdateLockerDetails(db, order.details());
-            }
-            var orderInserted = insertOrder(db, order, vendorId);
-            order.products().stream()
-                    .filter(product -> product.attachments() != null && product.attachments().size() > 1)
-                    .forEach(product -> {
-                                System.out.printf("Product %s in order %s of vendor %s has attachments.", product.part_number_key(), order.id(), order.vendor_name());
-                                reportIssue(order);
-                            }
-                    );
-            if (orderInserted.inserted) {
-                insertOrderDependents(db, order, orderInserted.surrogateId);
-            } else {
-                var oldOrder = selectWholeOrderResult(db, orderInserted.surrogateId, vendorName);
-                var hasDifferences = oldOrder.reportUnhandledDifferences(order);
-                if (hasDifferences) {
-                    System.out.printf("There are changes in order %s of vendor %s that are not handled yet.", order.id(), order.vendor_name());
-                    reportIssue(order);
-                }
-                if (order.modified() != null) {
-                    if ((oldOrder.modified() == null) || order.modified().isAfter(oldOrder.modified())) {
-                        updateTimestamp(db, orderInserted.surrogateId, "modified", order.modified());
-                    } else if (order.modified().isBefore(oldOrder.modified())) {
-                        System.out.printf("%s:%s -> %s:%s Modified date changed to an older date, from %s to %s.%n", oldOrder.vendor_name(), oldOrder.id(), order.vendor_name(), order.id(), oldOrder.modified(), order.modified());
-                    }
-                }
-                if (oldOrder.status() != order.status()) {
-                    System.out.printf("Update status for order %s, was %d will be %d%n.", order.id(), oldOrder.status(), order.status());
-                    updateInt(db, orderInserted.surrogateId, "status", order.status());
-                }
-                if (!Objects.equals(oldOrder.is_complete(), order.is_complete())) {
-                    updateInt(db, orderInserted.surrogateId, "is_complete", order.status());
-                }
-                if (!Objects.equals(oldOrder.late_shipment(), order.late_shipment())) {
-                    updateInt(db, orderInserted.surrogateId, "late_shipment", order.late_shipment());
-                }
-                if (!Objects.equals(oldOrder.payment_status(), order.payment_status())) {
-                    updateInt(db, orderInserted.surrogateId, "payment_status", order.payment_status());
-                }
-                if (!Objects.equals(oldOrder.reason_cancellation(), order.reason_cancellation())) {
-                    updateCancellationReason(db, orderInserted.surrogateId, order.reason_cancellation());
-                }
-                if (!Objects.equals(order.maximum_date_for_shipment(), oldOrder.maximum_date_for_shipment())) {
-                    updateTimestamp(db, orderInserted.surrogateId, "maximum_date_for_shipment", order.maximum_date_for_shipment());
-                }
-                if (!Objects.equals(oldOrder.finalization_date(), order.finalization_date())) {
-                    updateTimestamp(db, orderInserted.surrogateId, "finalization_date", order.finalization_date());
-                }
-                if (!Objects.equals(oldOrder.cashed_co(), order.cashed_co())) {
-                    updateNumeric(db, orderInserted.surrogateId, "cashed_co", order.cashed_co());
-                }
-                if (!Objects.equals(oldOrder.cashed_cod(), order.cashed_cod())) {
-                    updateNumeric(db, orderInserted.surrogateId, "cashed_cod", order.cashed_cod());
-                }
-                if (!Objects.equals(oldOrder.refunded_amount(), order.refunded_amount())) {
-                    updateNumeric(db, orderInserted.surrogateId, "refunded_amount", order.refunded_amount());
-                }
-                if (!Objects.equals(oldOrder.refund_status(), order.refund_status())) {
-                    updateString(db, orderInserted.surrogateId, "refund_status", order.refund_status());
-                }
-                if (!Objects.equals(oldOrder.delivery_mode(), order.delivery_mode())) {
-                    updateString(db, orderInserted.surrogateId, "delivery_mode", order.delivery_mode());
-                }
-                if (!Objects.equals(oldOrder.delivery_payment_mode(), order.delivery_payment_mode())) {
-                    updateString(db, orderInserted.surrogateId, "delivery_payment_mode", order.delivery_payment_mode());
-                }
-                if (!Objects.equals(oldOrder.payment_mode(), order.payment_mode())) {
-                    updateString(db, orderInserted.surrogateId, "payment_mode", order.payment_mode());
-                }
-                if (!Objects.equals(oldOrder.payment_mode_id(), order.payment_mode_id())) {
-                    updateInt(db, orderInserted.surrogateId, "payment_mode_id", order.payment_mode_id());
-                }
-                if (!Objects.equals(oldOrder.detailed_payment_method(), order.detailed_payment_method())) {
-                    updateString(db, orderInserted.surrogateId, "detailed_payment_method", order.detailed_payment_method());
-                }
-                updateOrderDependents(db, order, oldOrder, orderInserted.surrogateId);
-            }
-            return 0;
+            var vendorId = insertOrUpdateVendor(db, order.vendor_name(), account);
+            return addOrderResult(order, db, vendorId, order.vendor_name());
         });
-    }
-
-    private static final Scanner scanner = new Scanner(System.in);
-
-    private void reportIssue(OrderResult order) {
-        System.out.println("The above message indicates that for some changes, there is no code to handle them.");
-        System.out.println("A fix to the program is needed. You need to report this error to the developer.");
-        System.out.printf("Please report the vendor '%s' and order ID '%s'.%n", order.vendor_name(), order.id());
-        System.out.println("If you want, you can continue with the program or wait until a fix is provided.");
-        System.out.println("Please type the word “Yes” if you want the program to continue and fix the problem some other time. Type “No” if you want the program to stop and wait until you receive a fixed version of the program.");
-        System.out.println("If you type anything else, the program will stop now.");
-        String input = scanner.nextLine().trim().toLowerCase();
-        boolean isYes = input.equals("yes");
-        System.out.println("You entered: " + (isYes ? "YES, and the program will continue" : "NO, thus the program will terminate now."));
-        if (!isYes) System.exit(1);
-    }
-
-    /**
-     * Update a numeric field in the emag_order table.
-     *
-     * @param db database connection.
-     * @param surrogateId row id.
-     * @param field name of the field.
-     * @param newValue value to store.
-     * @return 1 if the row was updated or 0 if no row matched the surrogateId.
-     * @throws SQLException on database errors.
-     */
-    private static int updateNumeric(Connection db, int surrogateId, final String field, BigDecimal newValue) throws SQLException {
-        try (var s = db.prepareStatement("UPDATE emag_order SET " + field + " =? WHERE surrogate_id= ?")) {
-            s.setObject(1, newValue);
-            s.setInt(2, surrogateId);
-            return s.executeUpdate();
-        }
-    }
-
-    /**
-     * Update an integer field in the emag_order table.
-     *
-     * @param db database connection.
-     * @param surrogateId row id.
-     * @param field name of the field.
-     * @param newValue value to store.
-     * @return 1 if the row was updated or 0 if no row matched the surrogateId.
-     * @throws SQLException on database errors.
-     */
-    private static int updateInt(Connection db, int surrogateId, final String field, int newValue) throws SQLException {
-        try (var s = db.prepareStatement("UPDATE emag_order SET " + field + " =? WHERE surrogate_id= ?")) {
-            s.setInt(1, newValue);
-            s.setInt(2, surrogateId);
-            return s.executeUpdate();
-        }
-    }
-
-    /**
-     * Update a string field in the emag_order table.
-     *
-     * @param db database connection.
-     * @param surrogateId row id.
-     * @param field name of the field.
-     * @param newValue value to store.
-     * @return 1 if the row was updated or 0 if no row matched the surrogateId.
-     * @throws SQLException on database errors.
-     */
-    private static int updateString(Connection db, int surrogateId, final String field, String newValue) throws SQLException {
-        try (var s = db.prepareStatement("UPDATE emag_order SET " + field + " =? WHERE surrogate_id= ?")) {
-            s.setString(1, newValue);
-            s.setInt(2, surrogateId);
-            return s.executeUpdate();
-        }
-    }
-
-    /**
-     * Update a timestamp field in the emag_order table.
-     *
-     * @param db database connection.
-     * @param surrogateId row id.
-     * @param field name of the field.
-     * @param newValue value to store.
-     * @return 1 if the row was updated or 0 if no row matched the surrogateId.
-     * @throws SQLException on database errors.
-     */
-    private static int updateTimestamp(Connection db, int surrogateId, String field, LocalDateTime newValue) throws SQLException {
-        try (var s = db.prepareStatement("UPDATE emag_order SET " + field + " =? WHERE surrogate_id= ?")) {
-            s.setTimestamp(1, toTimestamp(newValue));
-            s.setInt(2, surrogateId);
-            return s.executeUpdate();
-        }
-    }
-
-    /**
-     * Update the cancellation reason in the emag_order table.
-     * This updates both cancellation_reason and cancellation_reason_text.
-     *
-     * @param db database connection.
-     * @param surrogateId row id.
-     * @param newValue value to store.
-     * @return 1 if the row was updated or 0 if no row matched the surrogateId.
-     * @throws SQLException on database errors.
-     */
-    private static int updateCancellationReason(Connection db, int surrogateId, CancellationReason newValue) throws SQLException {
-        try (var s = db.prepareStatement("UPDATE emag_order SET cancellation_reason = ?, cancellation_reason_text = ?  WHERE surrogate_id = ?")) {
-            s.setInt(1, newValue.id());
-            s.setString(2, newValue.name());
-            s.setInt(3, surrogateId);
-            return s.executeUpdate();
-        }
-    }
-
-    /**
-     * Update order products, flags, attachments and vouchers.
-     *
-     * @param db database connection.
-     * @param order order with the up-to-date values.
-     * @param oldOrder order as it was found in the database.
-     * @param surrogateId row id.
-     * @throws SQLException on database errors.
-     */
-    private void updateOrderDependents(Connection db, OrderResult order, OrderResult oldOrder, int surrogateId) throws SQLException {
-        if (!Objects.equals(oldOrder.flags(), order.flags())) {
-            updateFlags(db, order, surrogateId);
-        }
-        if (!Objects.equals(oldOrder.products(), order.products())) {
-            updateProducts(db, order, surrogateId);
-        }
-        if (!Objects.equals(oldOrder.attachments(), order.attachments())) {
-            updateAttachments(db, order, surrogateId);
-        }
-        if (!Objects.equals(oldOrder.vouchers(), order.vouchers())) {
-            updateVouchers(db, order, surrogateId);
-        }
-    }
-
-    private static void insertOrderDependents(Connection db, OrderResult order, int surrogateId) throws SQLException {
-        insertProducts(db, order, surrogateId);
-        if (order.shipping_tax_voucher_split() != null) {
-            for (var voucherSplit : order.shipping_tax_voucher_split()) {
-                insertVoucherSplit(db, voucherSplit, surrogateId);
-            }
-        }
-        insertAttachments(db, order, surrogateId);
-        insertVouchers(db, order, surrogateId);
-        insertFlags(db, order, surrogateId);
-        if (order.enforced_vendor_courier_accounts() != null) {
-            for (String enforced_vendor_courier_account : order.enforced_vendor_courier_accounts()) {
-                insertEnforcedVendorCourierAccount(db, enforced_vendor_courier_account, surrogateId);
-            }
-        }
-    }
-
-    private static void insertVouchers(Connection db, OrderResult order, int surrogateId) throws SQLException {
-        if (order.vouchers() != null) {
-            for (Voucher voucher : order.vouchers()) {
-                insertVoucher(db, voucher, surrogateId);
-            }
-        }
-    }
-
-    private static void insertAttachments(Connection db, OrderResult order, int surrogateId) throws SQLException {
-        if (order.attachments() != null) {
-            for (var attachment : order.attachments()) {
-                insertAttachment(db, attachment, surrogateId);
-            }
-        }
-    }
-
-    private static void insertProducts(Connection db, OrderResult order, int surrogateId) throws SQLException {
-        if (order.products() != null) {
-            for (var product : order.products()) {
-                insertProductInOrder(db, product, surrogateId);
-                insertVoucherSplits(db, surrogateId, product);
-            }
-        }
-    }
-
-    private static int deleteProducts(Connection db, int surrogateId) throws SQLException {
-        try (var s = db.prepareStatement("DELETE FROM product_in_order WHERE emag_order_surrogate_id = ?")) {
-            s.setInt(1, surrogateId);
-            return s.executeUpdate();
-        }
-    }
-
-    private static void updateProducts(Connection db, OrderResult order, int surrogateId) throws SQLException {
-        deleteVoucherSplits(db, surrogateId);
-        deleteProducts(db, surrogateId);
-        if (order.products() != null) {
-            for (var product : order.products()) {
-                insertProductInOrder(db, product, surrogateId);
-                insertVoucherSplits(db, surrogateId, product);
-            }
-        }
-    }
-
-    private static void insertVoucherSplits(Connection db, int surrogateId, Product product) throws SQLException {
-        for (var voucherSplit : product.product_voucher_split()) {
-            insertVoucherSplit(db, voucherSplit, surrogateId, product.id());
-        }
-    }
-
-    private static int deleteVoucherSplits(Connection db, int surrogateId) throws SQLException {
-        try (var s = db.prepareStatement("DELETE FROM voucher_split WHERE emag_order_surrogate_id = ?")) {
-            s.setInt(1, surrogateId);
-            return s.executeUpdate();
-        }
-    }
-
-    private static void insertFlags(Connection db, OrderResult order, int surrogateId) throws SQLException {
-        if (order.flags() != null) {
-            for (Flag flag : order.flags()) {
-                insertFlag(db, flag, surrogateId);
-            }
-        }
-    }
-
-    private static int deleteFlags(Connection db, int surrogateId) throws SQLException {
-        try (var s = db.prepareStatement("DELETE FROM flag WHERE emag_order_surrogate_id = ?")) {
-            s.setInt(1, surrogateId);
-            return s.executeUpdate();
-        }
-    }
-
-    private static int deleteAttachments(Connection db, int surrogateId) throws SQLException {
-        try (var s = db.prepareStatement("DELETE FROM attachment WHERE emag_order_surrogate_id = ?")) {
-            s.setInt(1, surrogateId);
-            return s.executeUpdate();
-        }
-    }
-
-    private static int deleteVouchers(Connection db, int surrogateId) throws SQLException {
-        try (var s = db.prepareStatement("DELETE FROM voucher WHERE emag_order_surrogate_id = ?")) {
-            s.setInt(1, surrogateId);
-            return s.executeUpdate();
-        }
-    }
-
-    private static void updateFlags(Connection db, OrderResult order, int surrogateId) throws SQLException {
-        deleteFlags(db, surrogateId);
-        insertFlags(db, order, surrogateId);
-    }
-
-    private static void updateAttachments(Connection db, OrderResult order, int surrogateId) throws SQLException {
-        deleteAttachments(db, surrogateId);
-        insertAttachments(db, order, surrogateId);
-    }
-
-    private static void updateVouchers(Connection db, OrderResult order, int surrogateId) throws SQLException {
-        deleteVouchers(db, surrogateId);
-        insertVouchers(db, order, surrogateId);
     }
 
     /**
@@ -468,6 +132,53 @@ public class EmagMirrorDB {
     }
 
     /**
+     * Reads orders from the database based on the specified order ID.
+     *
+     * @param orderId the unique identifier of the order to retrieve.
+     * @return a list of {@code OrderRecord} instances matching the given order ID.
+     * @throws SQLException if a database error occurs during the process.
+     */
+    public List<OrderRecord> readOrdersById(String orderId) throws SQLException {
+        return database.readTX(db -> OrderRecord.getOrdersById(db, orderId));
+    }
+
+    /**
+     * Reads orders from the database based on the specified order ID.
+     *
+     * @param orderId the unique identifier of the order to retrieve.
+     * @return a list of {@code OrderRecord} instances matching the given order ID.
+     * @throws SQLException if a database error occurs during the process.
+     */
+    public List<ProductInOrderRecord> readProductInOrderByOrderId(String orderId) throws SQLException {
+        return database.readTX(db -> ProductInOrderRecord.getProductInOrderByOrder(db, orderId));
+    }
+
+    /**
+     * Reads orders from made by the customer specified by its customer ID.
+     *
+     * @param customerId the unique identifier of the customer.
+     * @return a list of {@code OrderRecord} instances matching the given order ID.
+     * @throws SQLException if a database error occurs during the process.
+     */
+    public List<OrderRecord> readOrdersByCustomerId(int customerId) throws SQLException {
+        return database.readTX(db -> OrderRecord.getOrdersByCustomerId(db, customerId));
+    }
+
+    /**
+     * Reads customer specified by its customer ID.
+     *
+     * @param customerId the unique identifier of the customer.
+     * @return a list of {@code OrderRecord} instances matching the given order ID.
+     * @throws SQLException if a database error occurs during the process.
+     */
+    public List<CustomerRecord> readCustomerById(int customerId) throws SQLException {
+        return database.readTX(db -> CustomerRecord.getCustomerById(db, customerId));
+    }
+
+    public List<EmagFetchHistogram> readFetchHistogram() throws SQLException {
+        return database.readTX(EmagFetchLog::getFetchHistogram);
+    }
+    /**
      * Read orders in a format that is useful to update the employee sheets.
      *
      * @param pnk Product identifier.
@@ -480,105 +191,16 @@ public class EmagMirrorDB {
         return database.readTX(db -> getOrderDataByProductAndTime(db, pnk, startTime, endTime));
     }
 
-
-    private OrderResult selectWholeOrderResult(Connection db, int surrogateId, String vendorName) throws SQLException {
-        var customer = selectCustomerByOrderId(db, surrogateId);
-        var productIds = selectProductIdByOrderId(db, surrogateId);
-        var products = productIds.stream().map(productId -> {
-            List<VoucherSplit> voucherSplits;
-            try {
-                voucherSplits = selectVoucherSplitsByProductId(db, productId);
-                try {
-                    return selectProduct(db, productId, voucherSplits, new ArrayList<>());
-                } catch (SQLException e) {
-                    throw new RuntimeException("Error retrieving product with productId " + productId, e);
-                }
-            } catch (SQLException e) {
-                throw new RuntimeException("Error retrieving the voucher splits of the product with the productId %s.".formatted(productId), e);
-            }
-        }).toList();
-        var voucherSplits = selectVoucherSplitsByOrderId(db, surrogateId);
-        List<Attachment> attachments = selectAttachmentsByOrderId(db, surrogateId);
-        List<Voucher> vouchers = selectVouchersByOrderId(db, surrogateId);
-        var enforcedVendorCourierAccounts = selectEnforcedVendorCourierAccountsByOrderId(db, surrogateId);
-        var flags = selectFlagsByOrderId(db, surrogateId);
-        return selectOrder(db,
-                surrogateId,
-                vendorName,
-                customer,
-                voucherSplits,
-                products,
-                attachments,
-                vouchers,
-                enforcedVendorCourierAccounts,
-                flags);
-    }
-
-
-    private static void insertOrUpdateLockerDetails(Connection db, final LockerDetails details) throws SQLException {
-        var added = insertLockerDetails(db, details);
-        if (added == 0) {
-            var current = selectLockerDetails(db, details.locker_id());
-            if (!details.equals(current)) {
-                logger.log(FINE, () -> "LockerDetails differs:%n old: %s%n new: %s%n".formatted(current, details));
-                updateLockerDetails(db, details);
-            }
-        }
-    }
-
     /**
-     * Insert a customer into the database.
-     * If there is already a customer by the same ID, the record is updated if the data differs and the new record has
-     * a more recent modified date.
+     * Adds an RMA record to the database.
      *
-     * @param db database
-     * @param customer new customer record
-     * @throws SQLException in case of malfunction
+     * @param rmaResult the RMA details to be added.
+     * @throws SQLException if an error occurs during the database operation.
      */
-    private static void insertOrUpdateCustomer(Connection db, final Customer customer) throws SQLException {
-        var added = insertCustomer(db, customer);
-        if (added == 0) {
-            var current = selectCustomer(db, customer.id());
-            if (!customer.sameExceptForDate(current) && customer.modified().isAfter(current.modified())) {
-                logger.log(FINE, () -> "Updating customer:%n old: %s%n new: %s%n".formatted(current, customer));
-                updateCustomer(db, customer);
-            }
-        }
+    public void addRMA(RMAResult rmaResult) throws SQLException {
+        database.writeTX(db -> addRMAResult(db, rmaResult));
     }
 
-    public void addRMA(RMAResult rmaResult) throws SQLException {
-        database.writeTX(db -> {
-            var emagId = rmaResult.emag_id();
-            insertRMAResult(db, rmaResult);
-            if (rmaResult.products() != null) {
-                for (var product : rmaResult.products()) {
-                    insertReturnedProduct(db, product, emagId);
-                }
-            }
-            if (rmaResult.awbs() != null) {
-                for (var awb : rmaResult.awbs()) {
-                    insertAWB(db, awb, emagId);
-                }
-            }
-            if (rmaResult.request_history() != null) {
-                for (var requestHistory : rmaResult.request_history()) {
-                    insertRequestHistory(db, requestHistory, emagId);
-                }
-            }
-            if (rmaResult.status_history() != null) {
-                for (var statusHistory : rmaResult.status_history()) {
-                    var historyUUID = UUID.randomUUID();
-                    insertStatusHistory(db, statusHistory, emagId, historyUUID);
-                    if (statusHistory.requests() != null) {
-                        for (var statusRequest : statusHistory.requests()) {
-                            insertStatusRequest(db, statusRequest, historyUUID);
-                        }
-                    }
-                }
-            }
-            return 0;
-        });
-    }
 
     public void addOrUpdateProduct(ProductInfo productInfo) throws SQLException {
         database.writeTX(db -> {
@@ -677,68 +299,6 @@ public class EmagMirrorDB {
         );
     }
 
-    private Map<String, List<POInfo>> getOrderDataByProduct(Connection db, String productCode) throws SQLException {
-        var result = new HashMap<String, List<POInfo>>();
-        try (var s = db.prepareStatement("""
-                SELECT p.name AS productName,
-                pio.quantity AS quantity,
-                pio.initial_qty AS initialQuantity,
-                pio.storno_qty AS stornoQuantity,
-                pio.sale_price AS salePrice,
-                pio.vat AS vat,
-                pio.created, pio.modified,
-                o.status AS orderStatus,
-                o.date AS orderDate,
-                o.modified,
-                o.id AS orderId
-                FROM product_in_order AS pio
-                INNER JOIN product AS p ON p.emag_pnk = pio.part_number_key
-                INNER JOIN emag_order AS o ON pio.emag_order_surrogate_id = o.surrogate_id
-                WHERE p.product_code = ? AND (o.status = 4 OR o.status = 5)
-                ORDER BY o.id, o.status
-                """)) {
-            s.setObject(1, productCode);
-            try (var rs = s.executeQuery()) {
-                while (rs.next()) {
-                    var orderId = rs.getString("orderId");
-                    var poInfos = result.getOrDefault(orderId, new ArrayList<>());
-                    var quantity = rs.getInt("quantity");
-                    var initialQuantity = rs.getInt("initialQuantity");
-                    var stornoQuantity = rs.getInt("stornoQuantity");
-                    var salePrice = rs.getBigDecimal("salePrice");
-                    var vat = new BigDecimal(rs.getString("vat"));
-                    var price = vat.add(BigDecimal.ONE).multiply(salePrice).setScale(2, HALF_EVEN);
-                    var poInfo = new POInfo(
-                            orderId,
-                            toLocalDate(rs.getTimestamp("orderDate")),
-                            rs.getInt("orderStatus"),
-                            rs.getString("productName"),
-                            quantity,
-                            initialQuantity,
-                            stornoQuantity,
-                            price
-                    );
-                    poInfos.add(poInfo);
-                    result.put(orderId, poInfos);
-                }
-
-            }
-        }
-        return result;
-    }
-
-    private List<String> getProductCodes(Connection db) throws SQLException {
-        var products = new ArrayList<String>();
-        try (var s = db.prepareStatement("SELECT product_code FROM product")) {
-            try (var rs = s.executeQuery()) {
-                while (rs.next()) {
-                    products.add(rs.getString(1));
-                }
-            }
-        }
-        return products;
-    }
-
     public List<ProductInfo> readProducts() throws SQLException {
         return database.readTX(ProductTable::getProducts);
     }
@@ -749,44 +309,7 @@ public class EmagMirrorDB {
      * @throws SQLException on database errors.
      */
     public void updateGMVTable() throws SQLException {
-        database.writeTX(db -> {
-                    var products = getProductCodes(db);
-                    for (String productCode : products) {
-                        var ordersWithProduct = getOrderDataByProduct(db, productCode);
-                        var gmvByMonth = new HashMap<YearMonth, BigDecimal>();
-                        ordersWithProduct.forEach((_, poInfos) -> {
-                            if (poInfos.size() == 1) {
-                                POInfo order = poInfos.getFirst();
-                                var yearMonth = YearMonth.from(order.orderDate);
-                                addToGMV(gmvByMonth, yearMonth, order);
-                            } else if (poInfos.size() == 2) {
-                                POInfo finalized = poInfos.getFirst();
-                                POInfo storno = poInfos.getLast();
-                                // Two entries for the same product in the same order.
-                                if (storno.orderStatus == 4 || finalized.orderStatus == 5) {
-                                    specialCase(gmvByMonth, finalized, storno);
-                                } else {
-                                    if (storno.initialQuantity != finalized.quantity) {
-                                        logger.log(WARNING, "Mismatch in quantity between\n finalzed order: %s\n and storno order: %s.".formatted(finalized, storno));
-                                    }
-                                    var finalizedMonth = YearMonth.from(finalized.orderDate);
-                                    var stornoMonth = YearMonth.from(storno.orderDate);
-                                    if (finalizedMonth.equals(stornoMonth)) {
-                                        addToGMV(gmvByMonth, finalizedMonth, storno);
-                                    } else {
-                                        addToGMV(gmvByMonth, finalizedMonth, finalized);
-                                        subtractFromGMV(gmvByMonth, stornoMonth, storno);
-                                    }
-                                }
-                            } else {
-                                throw new RuntimeException("Not prepared to handle order with more than two states: %s.".formatted(poInfos));
-                            }
-                        });
-                        insertOrUpdateGMV(db, productCode, gmvByMonth);
-                    }
-                    return true;
-                }
-        );
+        database.writeTX(EmagMirrorDB::computeGMV);
     }
 
     /**
@@ -800,6 +323,10 @@ public class EmagMirrorDB {
         return database.readTX(db -> getGMVByMonth(db, month));
     }
 
+    public SortedMap<ProductInfo, SortedMap<YearMonth, BigDecimal>> getGMVTable() throws SQLException {
+        return database.readTX(EmagMirrorDB::getGMV);
+    }
+
     /**
      * Retrieve the full GMV table.
      *
@@ -807,7 +334,7 @@ public class EmagMirrorDB {
      * @return map from productInfo to map of month to GMV.
      * @throws SQLException on database error
      */
-    public SortedMap<ProductInfo, SortedMap<YearMonth, BigDecimal>> getGMV(Connection db) throws SQLException {
+    public static SortedMap<ProductInfo, SortedMap<YearMonth, BigDecimal>> getGMV(Connection db) throws SQLException {
         var rows = new TreeMap<ProductInfo, SortedMap<YearMonth, BigDecimal>>(
                 Comparator.comparing(ProductInfo::name)
         );
@@ -873,7 +400,6 @@ public class EmagMirrorDB {
                 s.setInt(1, year);
                 try (var rs = s.executeQuery()) {
                     while (rs.next()) {
-// 2024-10-17 16:49:51
                         var priceWithoutVAT = rs.getBigDecimal(6);
                         var priceWithVAT = priceWithoutVAT.multiply(BigDecimal.valueOf(1.19)); // TODO: Proper handling of VAT required
                         String customerName = rs.getString(8);
@@ -923,78 +449,6 @@ public class EmagMirrorDB {
             }
             return rows;
         });
-    }
-
-    private List<EmployeeSheetData> getOrderDataByProductAndTime(Connection db, String pnk, LocalDateTime startTime, LocalDateTime endTime) throws SQLException {
-        var list = new ArrayList<EmployeeSheetData>();
-        try (var s = db.prepareStatement(
-                //language=postgres-sql
-                """
-                        SELECT
-                          o.id,
-                          p.quantity,
-                          p.sale_price,
-                          c.legal_entity,
-                          o.date,
-                          pi.name AS product_name,
-                          p.part_number_key,
-                          c.name,
-                          c.billing_name,
-                          c.billing_phone,
-                          CONCAT_WS(', ', c.billing_locality_id, c.billing_street, c.billing_country, c.billing_postal_code, c.billing_suburb, c.billing_city) AS billing_address,
-                          c.name AS customer_name,
-                          c.shipping_phone,
-                          CONCAT_WS(', ', c.shipping_locality_id, c.shipping_street, c.shipping_country, c.shipping_postal_code, c.shipping_suburb, c.shipping_city) AS shipping_address,
-                          o.delivery_mode,
-                          p.currency,
-                          p.vat,
-                          o.status
-                        FROM emag_order as o
-                        LEFT JOIN customer as c
-                        ON o.customer_id = c.id
-                        LEFT JOIN vendor as v
-                        ON o.vendor_id = v.id
-                        INNER JOIN product_in_order as p
-                        ON p.emag_order_surrogate_id = o.surrogate_id
-                        LEFT JOIN product as pi
-                        ON p.part_number_key = pi.emag_pnk
-                        WHERE (o.status = 4 OR o.status = 5) AND p.part_number_key = ? AND ? < o.date AND o.date < ?
-                        ORDER BY o.date
-                        """)) {
-            s.setString(1, pnk);
-            s.setTimestamp(2, toTimestamp(startTime));
-            s.setTimestamp(3, toTimestamp(endTime));
-            try (var rs = s.executeQuery()) {
-                while (rs.next()) {
-                    var priceWithoutVAT = rs.getBigDecimal("sale_price");
-                    var vatRate = new BigDecimal(rs.getString("vat"));
-                    var vat = priceWithoutVAT.multiply(vatRate);
-                    var priceWithVAT = priceWithoutVAT.add(vat);
-                    int status = rs.getInt("status");
-                    var data = new EmployeeSheetData(
-                            rs.getString("id"),
-                            rs.getInt("quantity"),
-                            priceWithVAT,
-                            rs.getInt("legal_entity")>0,
-                            toLocalDateTime(rs.getTimestamp("date")),
-                            rs.getString("product_name"),
-                            rs.getString("part_number_key"),
-                            rs.getString("customer_name"),
-                            rs.getString("billing_name"),
-                            rs.getString("billing_phone"),
-                            rs.getString("billing_address"),
-                            rs.getString("customer_name"),
-                            rs.getString("shipping_phone"),
-                            rs.getString("shipping_address"),
-                            rs.getString("delivery_mode"),
-                            false
-                    );
-                    list.add(data);
-
-                }
-            }
-        }
-        return list;
     }
 
     /**
@@ -1108,7 +562,6 @@ public class EmagMirrorDB {
                             """)) {
                 try (var rs = s.executeQuery()) {
                     while (rs.next()) {
-// 2024-10-17 16:49:51
                         var priceWithoutVAT = rs.getBigDecimal(6);
                         var priceWithVAT = priceWithoutVAT.multiply(BigDecimal.valueOf(1.19)); // TODO: Proper handling of VAT required
                         String customerName = rs.getString(8);
@@ -1173,860 +626,163 @@ public class EmagMirrorDB {
         });
     }
 
-    /**
-     * Map the vendor name found in the eMAG order to a UUID we can use in the database.
-     * It will add an entry whenever a new vendor name appears.
-     *
-     * @param name of vendor.
-     * @return UUID associated with the vendor.
-     * @throws SQLException on database errors.
-     */
-    private UUID getOrAddVendor(String name, String account) throws SQLException {
-        return database.writeTX(db -> {
-            UUID id = selectVendorIdByName(db, name);
-            if (id == null) {
-                id = UUID.randomUUID();
-                try (var s = db.prepareStatement("INSERT INTO vendor (id, vendor_name, isfbe, account) VALUES (?,?,?,?)")) {
-                    s.setObject(1, id);
-                    s.setString(2, name);
-                    // Zoopie Invest is a special case, it does not contain FBE in the name, but is FBE.
-                    s.setBoolean(3, name.contains("FBE") || name.contains("Zoopie Invest"));
-                    s.setString(4, account);
-                    s.executeUpdate();
-                }
-            }
-            return id;
-        });
-    }
-
-
-    private static int insertAttachment(Connection db, Attachment attachment, int surrogateId) throws SQLException {
-        try (var s = db.prepareStatement("INSERT INTO attachment (emag_order_surrogate_id, name, url, type, force_download, visibility) VALUES (?, ?, ?, ?, ?, ?) ON CONFLICT(emag_order_surrogate_id, url) DO NOTHING")) {
-            s.setInt(1, surrogateId);
-            s.setString(2, attachment.name());
-            s.setString(3, attachment.url());
-            s.setInt(4, attachment.type());
-            s.setInt(5, attachment.force_download());
-            s.setString(6, attachment.visibility());
-            return s.executeUpdate();
-        }
-    }
-
-    private List<Attachment> selectAttachmentsByOrderId(Connection db, int surrogateId) throws SQLException {
-        String query = "SELECT name, url, type, force_download, visibility FROM attachment WHERE emag_order_surrogate_id = ?";
-        var list = new ArrayList<Attachment>();
-        try (var s = db.prepareStatement(query)) {
-            s.setInt(1, surrogateId);
-            try (var rs = s.executeQuery()) {
-                while (rs.next()) {
-                    list.add(new Attachment(rs.getString("name"), rs.getString("url"), rs.getInt("type"), rs.getInt("force_download"), rs.getString("visibility")));
-                }
-            }
-        }
-        return list;
-    }
-
-    private static int insertFlag(Connection db, Flag flag, int surrogateId) throws SQLException {
-        try (var s = db.prepareStatement("INSERT INTO flag (emag_order_surrogate_id, flag, value) VALUES (?, ?, ?) ON CONFLICT(emag_order_surrogate_id, flag) DO NOTHING")) {
-            s.setInt(1, surrogateId);
-            s.setString(2, flag.flag());
-            s.setString(3, flag.value());
-            return s.executeUpdate();
-        }
-    }
-
-    private List<Flag> selectFlagsByOrderId(Connection db, int surrogateId) throws SQLException {
-        var list = new ArrayList<Flag>();
-        String query = "SELECT emag_order_surrogate_id, flag, value FROM flag WHERE emag_order_surrogate_id = ?";
-        try (var s = db.prepareStatement(query)) {
-            s.setInt(1, surrogateId);
-            try (var rs = s.executeQuery()) {
-                if (rs.next()) {
-                    list.add(new Flag(
-                            rs.getString("flag"),
-                            rs.getString("value")
-                    ));
+    private static boolean computeGMV(Connection db) throws SQLException {
+        var products = getProductCodes(db);
+        for (String productCode : products) {
+            var ordersWithProduct = getOrderDataByProduct(db, productCode);
+            var gmvByMonth = new HashMap<YearMonth, BigDecimal>();
+            ordersWithProduct.forEach((_, poInfos) -> {
+                if (poInfos.size() == 1) {
+                    POInfo order = poInfos.getFirst();
+                    var yearMonth = YearMonth.from(order.orderDate);
+                    addToGMV(gmvByMonth, yearMonth, order);
+                } else if (poInfos.size() == 2) {
+                    POInfo finalized = poInfos.getFirst();
+                    POInfo storno = poInfos.getLast();
+                    // Two entries for the same product in the same order.
+                    if (storno.orderStatus == 4 || finalized.orderStatus == 5) {
+                        specialCase(gmvByMonth, finalized, storno);
+                    } else {
+                        if (storno.initialQuantity != finalized.quantity) {
+                            logger.log(WARNING, "Mismatch in quantity between\n finalzed order: %s\n and storno order: %s.".formatted(finalized, storno));
+                        }
+                        var finalizedMonth = YearMonth.from(finalized.orderDate);
+                        var stornoMonth = YearMonth.from(storno.orderDate);
+                        if (finalizedMonth.equals(stornoMonth)) {
+                            addToGMV(gmvByMonth, finalizedMonth, storno);
+                        } else {
+                            addToGMV(gmvByMonth, finalizedMonth, finalized);
+                            subtractFromGMV(gmvByMonth, stornoMonth, storno);
+                        }
+                    }
                 } else {
-                    return null; // or throw an exception if not found
+                    throw new RuntimeException("Not prepared to handle order with more than two states: %s.".formatted(poInfos));
                 }
-            }
+            });
+            insertOrUpdateGMV(db, productCode, gmvByMonth);
         }
-        return list;
-    }
-    private static int insertEnforcedVendorCourierAccount(Connection db, String enforcedVendorCourierAccount, int surrogateId) throws SQLException {
-        try (var s = db.prepareStatement("INSERT INTO enforced_vendor_courier_account (emag_order_surrogate_id, courier) VALUES (?, ?) ON CONFLICT(emag_order_surrogate_id, courier)")) {
-            s.setInt(1, surrogateId);
-            s.setString(2, enforcedVendorCourierAccount);
-            return s.executeUpdate();
-        }
+        return true;
     }
 
-    private static List<String> selectEnforcedVendorCourierAccountsByOrderId(Connection db, int surrogateId) throws SQLException {
-        var list = new ArrayList<String>();
-        String query = "SELECT courier FROM enforced_vendor_courier_account WHERE emag_order_surrogate_id = ?";
-        try (var s = db.prepareStatement(query)) {
-            s.setInt(1, surrogateId);
+    private List<EmployeeSheetData> getOrderDataByProductAndTime(Connection db, String pnk, LocalDateTime startTime, LocalDateTime endTime) throws SQLException {
+        var list = new ArrayList<EmployeeSheetData>();
+        try (var s = db.prepareStatement(
+                //language=postgres-sql
+                """
+                        SELECT
+                          o.id,
+                          p.quantity,
+                          p.sale_price,
+                          c.legal_entity,
+                          o.date,
+                          pi.name AS product_name,
+                          p.part_number_key,
+                          c.name,
+                          c.billing_name,
+                          c.billing_phone,
+                          CONCAT_WS(', ', c.billing_locality_id, c.billing_street, c.billing_country, c.billing_postal_code, c.billing_suburb, c.billing_city) AS billing_address,
+                          c.name AS customer_name,
+                          c.shipping_phone,
+                          CONCAT_WS(', ', c.shipping_locality_id, c.shipping_street, c.shipping_country, c.shipping_postal_code, c.shipping_suburb, c.shipping_city) AS shipping_address,
+                          o.delivery_mode,
+                          p.currency,
+                          p.vat,
+                          o.status
+                        FROM emag_order as o
+                        LEFT JOIN customer as c
+                        ON o.customer_id = c.id
+                        LEFT JOIN vendor as v
+                        ON o.vendor_id = v.id
+                        INNER JOIN product_in_order as p
+                        ON p.emag_order_surrogate_id = o.surrogate_id
+                        LEFT JOIN product as pi
+                        ON p.part_number_key = pi.emag_pnk
+                        WHERE (o.status = 4 OR o.status = 5) AND p.part_number_key = ? AND ? < o.date AND o.date < ?
+                        ORDER BY o.date
+                        """)) {
+            s.setString(1, pnk);
+            s.setTimestamp(2, toTimestamp(startTime));
+            s.setTimestamp(3, toTimestamp(endTime));
             try (var rs = s.executeQuery()) {
                 while (rs.next()) {
-                    list.add(rs.getString("courier"));
-                }
-            }
-        }
-        return list;
-    }
-    private static int insertVoucher(Connection db, Voucher voucher, int surrogateId) throws SQLException {
-        try (var s = db.prepareStatement("INSERT INTO voucher (voucher_id, emag_order_surrogate_id, modified, created, status, sale_price_vat, sale_price, voucher_name, vat, issue_date, id) VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?) ON CONFLICT(voucher_id) DO NOTHING")) {
-            s.setInt(1, voucher.voucher_id());
-            s.setInt(2, surrogateId);
-            s.setString(3, voucher.modified());
-            s.setString(4, voucher.created());
-            s.setInt(5, voucher.status());
-            s.setBigDecimal(6, voucher.sale_price_vat());
-            s.setBigDecimal(7, voucher.sale_price());
-            s.setString(8, voucher.voucher_name());
-            s.setBigDecimal(9, voucher.vat());
-            s.setString(10, voucher.issue_date());
-            s.setString(11, voucher.id());
-            return s.executeUpdate();
-        }
-    }
-
-    private List<Voucher> selectVouchersByOrderId(Connection db, int surrogateId) throws SQLException {
-        var list = new ArrayList<Voucher>();
-        String query = """
-                SELECT voucher_id, modified, created, status, sale_price_vat, sale_price, voucher_name, vat, issue_date, id
-                FROM voucher
-                WHERE emag_order_surrogate_id = ?
-                """;
-        try (var s = db.prepareStatement(query)) {
-            s.setInt(1, surrogateId);
-            try (var rs = s.executeQuery()) {
-                while (rs.next()) {
-                    list.add(new Voucher(rs.getInt("voucher_id"),
-                            rs.getString("modified"),
-                            rs.getString("created"),
-                            rs.getInt("status"),
-                            rs.getBigDecimal("sale_price_vat"),
-                            rs.getBigDecimal("sale_price"),
-                            rs.getString("voucher_name"),
-                            rs.getBigDecimal("vat"),
-                            rs.getString("issue_date"),
-                            rs.getString("id")));
-                }
-            }
-        }
-        return list;
-    }
-
-    private static int insertVoucherSplit(Connection conn, VoucherSplit voucherSplit, int surrogateId) throws SQLException {
-        try (var s = conn.prepareStatement("INSERT INTO voucher_split (voucher_id, emag_order_surrogate_id, value, vat_value) VALUES (?, ?, ?, ?) ON CONFLICT(voucher_id) DO NOTHING")) {
-            s.setInt(1, voucherSplit.voucher_id());
-            s.setInt(2, surrogateId);
-            s.setBigDecimal(3, voucherSplit.value());
-            s.setBigDecimal(4, voucherSplit.vat_value());
-            return s.executeUpdate();
-        }
-    }
-
-    private List<VoucherSplit> selectVoucherSplitsByProductId(Connection db, int productId) throws SQLException {
-        var list = new ArrayList<VoucherSplit>();
-        try (var s = db.prepareStatement("""
-                SELECT voucher_id, value, vat_value, vat, offered_by FROM voucher_split WHERE product_id = ?
-                """)) {
-            s.setInt(1, productId);
-            try (var rs = s.executeQuery()) {
-                while (rs.next()) {
-                    list.add(new VoucherSplit(rs.getObject("voucher_id", Integer.class),
-                            rs.getBigDecimal("value"),
-                            rs.getBigDecimal("vat_value"),
-                            rs.getString("vat"),
-                            rs.getString("offered_by")));
-                }
-            }
-        }
-        return list;
-    }
-
-    private List<VoucherSplit> selectVoucherSplitsByOrderId(Connection db, int surrogateId) throws SQLException {
-        var list = new ArrayList<VoucherSplit>();
-        try (var s = db.prepareStatement("""
-                SELECT voucher_id, value, vat_value, vat, offered_by FROM voucher_split WHERE emag_order_surrogate_id = ?
-                """)) {
-            s.setInt(1, surrogateId);
-            try (var rs = s.executeQuery()) {
-                while (rs.next()) {
-                    list.add(new VoucherSplit(rs.getObject("voucher_id", Integer.class),
-                            rs.getBigDecimal("value"),
-                            rs.getBigDecimal("vat_value"),
-                            rs.getString("vat"),
-                            rs.getString("offered_by")));
-                }
-            }
-        }
-        return list;
-    }
-
-
-    private static int insertVoucherSplit(Connection conn, VoucherSplit voucherSplit, int surrogateId, int productId) throws SQLException {
-        try (var s = conn.prepareStatement("INSERT INTO voucher_split (voucher_id, emag_order_surrogate_id, product_id, value, vat_value, vat, offered_by) VALUES (?, ?, ?, ?, ?, ?, ?) ON CONFLICT(voucher_id) DO NOTHING")) {
-            s.setInt(1, voucherSplit.voucher_id());
-            s.setInt(2, surrogateId);
-            s.setInt(3, productId);
-            s.setBigDecimal(4, voucherSplit.value());
-            s.setBigDecimal(5, voucherSplit.vat_value());
-            s.setString(6, voucherSplit.vat());
-            s.setString(7, voucherSplit.offered_by());
-            return s.executeUpdate();
-        }
-    }
-
-    private static int insertProductInOrder(Connection db, Product product, int surrogateId) throws SQLException {
-        try (var s = db.prepareStatement("INSERT INTO product_in_order (id, emag_order_surrogate_id, product_id, mkt_id, name, status, ext_part_number, part_number, part_number_key, currency, vat, retained_amount, quantity, initial_qty, storno_qty, reversible_vat_charging, sale_price, original_price, created, modified, details, recycle_warranties, serial_numbers) VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?) ON CONFLICT(id) DO NOTHING")) {
-            s.setInt(1, product.id());
-            s.setInt(2, surrogateId);
-            s.setInt(3, product.product_id());
-            s.setInt(4, product.mkt_id());
-            s.setString(5, product.name());
-            s.setInt(6, product.status());
-            s.setString(7, product.ext_part_number());
-            s.setString(8, product.part_number());
-            s.setString(9, product.part_number_key());
-            s.setString(10, product.currency());
-            s.setString(11, product.vat());
-            s.setInt(12, product.retained_amount());
-            s.setInt(13, product.quantity());
-            s.setInt(14, product.initial_qty());
-            s.setInt(15, product.storno_qty());
-            s.setInt(16, product.reversible_vat_charging());
-            s.setBigDecimal(17, product.sale_price());
-            s.setBigDecimal(18, product.original_price());
-            s.setTimestamp(19, toTimestamp(product.created()));
-            s.setTimestamp(20, toTimestamp(product.modified()));
-            s.setString(21, String.join("\n", product.details()));
-            s.setString(22, String.join("\n", product.recycle_warranties()));
-            s.setString(23, product.serial_numbers());
-            return s.executeUpdate();
-        }
-    }
-
-    /**
-     * Read all product ID belonging to an order.
-     *
-     * @param db database
-     * @param surrogateId the synthetic ID of the order
-     * @return list of product IDs.
-     * @throws SQLException on database error
-     */
-    private List<Integer> selectProductIdByOrderId(Connection db, int surrogateId) throws SQLException {
-        var list = new ArrayList<Integer>();
-        try (var s = db.prepareStatement("""
-                SELECT id
-                FROM product_in_order
-                WHERE emag_order_surrogate_id = ?
-                """)) {
-            s.setInt(1, surrogateId);
-            try (var rs = s.executeQuery()) {
-                while (rs.next()) {
-                    list.add(rs.getInt("id"));
-                }
-            }
-        }
-        return list;
-
-    }
-
-    private Product selectProduct(Connection db, int productId, List<VoucherSplit> product_voucher_splits, List<Attachment> attachments) throws SQLException {
-        Product product = null;
-        try (var s = db.prepareStatement("""
-                SELECT id, product_id, mkt_id, name, status, ext_part_number, part_number, part_number_key, currency, vat, retained_amount, quantity, initial_qty, storno_qty, reversible_vat_charging, sale_price, original_price, created, modified, details, recycle_warranties, serial_numbers
-                FROM product_in_order
-                WHERE id = ?
-                """)) {
-            s.setInt(1, productId);
-            try (var rs = s.executeQuery()) {
-                if (rs.next()) {
-                    product = new Product(rs.getInt("id"),
-                            rs.getInt("product_id"),
-                            rs.getInt("mkt_id"),
-                            rs.getString("name"),
-                            product_voucher_splits,
-                            rs.getInt("status"),
-                            rs.getString("ext_part_number"),
-                            rs.getString("part_number"),
-                            rs.getString("part_number_key"),
-                            rs.getString("currency"),
-                            rs.getString("vat"),
-                            rs.getInt("retained_amount"),
+                    var priceWithoutVAT = rs.getBigDecimal("sale_price");
+                    var vatRate = new BigDecimal(rs.getString("vat"));
+                    var vat = priceWithoutVAT.multiply(vatRate);
+                    var priceWithVAT = priceWithoutVAT.add(vat);
+                    int status = rs.getInt("status");
+                    var data = new EmployeeSheetData(
+                            rs.getString("id"),
                             rs.getInt("quantity"),
-                            rs.getInt("initial_qty"),
-                            rs.getInt("storno_qty"),
-                            rs.getInt("reversible_vat_charging"),
-                            rs.getBigDecimal("sale_price"),
-                            rs.getBigDecimal("original_price"),
-                            toLocalDateTime(rs.getTimestamp("created")),
-                            toLocalDateTime(rs.getTimestamp("modified")),
-                            Arrays.asList(rs.getString("details").split("\\n")),
-                            Arrays.asList(rs.getString("recycle_warranties").split("\\n")),
-                            attachments,
-                            rs.getString("serial_numbers"));
-                }
-                if (rs.next()) {
-                    throw new RuntimeException("More than one product with the same ID " + productId);
-                }
-            }
-        }
-        return product;
-    }
-
-    private static int insertCustomer(Connection db, Customer c) throws SQLException {
-        try (var s = db.prepareStatement("""
-                INSERT INTO customer (
-                id, mkt_id, name, email, company, gender, code, registration_number, bank, iban, fax, legal_entity,
-                is_vat_payer, phone_1, phone_2, phone_3, billing_name, billing_phone, billing_country, billing_suburb,
-                billing_city, billing_locality_id, billing_street, billing_postal_code, liable_person, shipping_country,
-                shipping_suburb, shipping_city, shipping_locality_id, shipping_street,
-                shipping_postal_code, shipping_contact, shipping_phone, created, modified
-                )
-                VALUES (?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?) ON CONFLICT(id) DO NOTHING""")) {
-            s.setInt(1, c.id());
-            s.setInt(2, c.mkt_id());
-            s.setString(3, c.name());
-            s.setString(4, c.email());
-            s.setString(5, c.company());
-            s.setString(6, c.gender());
-            s.setString(7, c.code());
-            s.setString(8, c.registration_number());
-            s.setString(9, c.bank());
-            s.setString(10, c.iban());
-            s.setString(11, c.fax());
-            s.setInt(12, c.legal_entity());
-            s.setInt(13, c.is_vat_payer());
-            s.setString(14, c.phone_1());
-            s.setString(15, c.phone_2());
-            s.setString(16, c.phone_3());
-            s.setString(17, c.billing_name());
-            s.setString(18, c.billing_phone());
-            s.setString(19, c.billing_country());
-            s.setString(20, c.billing_suburb());
-            s.setString(21, c.billing_city());
-            s.setString(22, c.billing_locality_id());
-            s.setString(23, c.billing_street());
-            s.setString(24, c.billing_postal_code());
-            s.setString(25, c.liable_person());
-            s.setString(26, c.shipping_country());
-            s.setString(27, c.shipping_suburb());
-            s.setString(28, c.shipping_city());
-            s.setString(29, c.shipping_locality_id());
-            s.setString(30, c.shipping_street());
-            s.setString(31, c.shipping_postal_code());
-            s.setString(32, c.shipping_contact());
-            s.setString(33, c.shipping_phone());
-            s.setTimestamp(34, toTimestamp(c.created()));
-            s.setTimestamp(35, toTimestamp(c.modified()));
-            return s.executeUpdate();
-        }
-    }
-
-    private static Customer selectCustomer(Connection db, int customerId) throws SQLException {
-        Customer customer = null;
-        try (var s = db.prepareStatement("SELECT * FROM customer WHERE id = ?")) {
-            s.setInt(1, customerId);
-            try (var rs = s.executeQuery()) {
-                if (rs.next()) {
-                    customer = new Customer(customerId, rs.getInt("mkt_id"), rs.getString("name"), rs.getString("email"), rs.getString("company"), rs.getString("gender"), rs.getString("code"), rs.getString("registration_number"), rs.getString("bank"), rs.getString("iban"), rs.getString("fax"), rs.getInt("legal_entity"), rs.getInt("is_vat_payer"), rs.getString("phone_1"), rs.getString("phone_2"), rs.getString("phone_3"), rs.getString("billing_name"), rs.getString("billing_phone"), rs.getString("billing_country"), rs.getString("billing_suburb"), rs.getString("billing_city"), rs.getString("billing_locality_id"), rs.getString("billing_street"), rs.getString("billing_postal_code"), rs.getString("liable_person"), rs.getString("shipping_country"), rs.getString("shipping_suburb"), rs.getString("shipping_city"), rs.getString("shipping_locality_id"), rs.getString("shipping_street"), rs.getString("shipping_postal_code"), rs.getString("shipping_contact"), rs.getString("shipping_phone"), toLocalDateTime(rs.getTimestamp("created")), toLocalDateTime(rs.getTimestamp("modified")));
-                }
-            }
-        }
-        return customer;
-    }
-
-    private static Customer selectCustomerByOrderId(Connection db, int surrogateId) throws SQLException {
-        Customer customer = null;
-        try (var s = db.prepareStatement("""
-                SELECT * FROM customer AS c
-                INNER JOIN emag_order as o
-                ON c.id = o.customer_id
-                WHERE o.surrogate_id = ?
-                """)) {
-            s.setInt(1, surrogateId);
-            try (var rs = s.executeQuery()) {
-                if (rs.next()) {
-                    customer = new Customer(rs.getInt("id"),
-                            rs.getInt("mkt_id"),
-                            rs.getString("name"),
-                            rs.getString("email"),
-                            rs.getString("company"),
-                            rs.getString("gender"),
-                            rs.getString("code"),
-                            rs.getString("registration_number"),
-                            rs.getString("bank"),
-                            rs.getString("iban"),
-                            rs.getString("fax"),
-                            rs.getInt("legal_entity"),
-                            rs.getInt("is_vat_payer"),
-                            rs.getString("phone_1"),
-                            rs.getString("phone_2"),
-                            rs.getString("phone_3"),
+                            priceWithVAT,
+                            rs.getInt("legal_entity")>0,
+                            toLocalDateTime(rs.getTimestamp("date")),
+                            rs.getString("product_name"),
+                            rs.getString("part_number_key"),
+                            rs.getString("customer_name"),
                             rs.getString("billing_name"),
                             rs.getString("billing_phone"),
-                            rs.getString("billing_country"),
-                            rs.getString("billing_suburb"),
-                            rs.getString("billing_city"),
-                            rs.getString("billing_locality_id"),
-                            rs.getString("billing_street"),
-                            rs.getString("billing_postal_code"),
-                            rs.getString("liable_person"),
-                            rs.getString("shipping_country"),
-                            rs.getString("shipping_suburb"),
-                            rs.getString("shipping_city"),
-                            rs.getString("shipping_locality_id"),
-                            rs.getString("shipping_street"),
-                            rs.getString("shipping_postal_code"),
-                            rs.getString("shipping_contact"),
+                            rs.getString("billing_address"),
+                            rs.getString("customer_name"),
                             rs.getString("shipping_phone"),
-                            toLocalDateTime(rs.getTimestamp("created")),
-                            toLocalDateTime(rs.getTimestamp("modified")));
+                            rs.getString("shipping_address"),
+                            rs.getString("delivery_mode"),
+                            false
+                    );
+                    list.add(data);
+
                 }
             }
         }
-        return customer;
+        return list;
     }
 
-
-    private static int updateCustomer(Connection db, Customer c) throws SQLException {
-        try (var s = db.prepareStatement("UPDATE customer SET mkt_id = ?, name = ?, email = ?, company = ?, gender = ?, code = ?, registration_number = ?, bank = ?, iban = ?, fax = ?, legal_entity = ?, is_vat_payer = ?, phone_1 = ?, phone_2 = ?, phone_3 = ?, billing_name = ?, billing_phone = ?, billing_country = ?, billing_suburb = ?, billing_city = ?, billing_locality_id = ?, billing_street = ?, billing_postal_code = ?, liable_person = ?, shipping_country = ?, shipping_suburb = ?, shipping_city = ?, shipping_locality_id = ?, shipping_street = ?, shipping_postal_code = ?, shipping_contact = ?, shipping_phone = ?, modified = ? WHERE id = ?")) {
-            s.setInt(1, c.mkt_id());
-            s.setString(2, c.name());
-            s.setString(3, c.email());
-            s.setString(4, c.company());
-            s.setString(5, c.gender());
-            s.setString(6, c.code());
-            s.setString(7, c.registration_number());
-            s.setString(8, c.bank());
-            s.setString(9, c.iban());
-            s.setString(10, c.fax());
-            s.setInt(11, c.legal_entity());
-            s.setInt(12, c.is_vat_payer());
-            s.setString(13, c.phone_1());
-            s.setString(14, c.phone_2());
-            s.setString(15, c.phone_3());
-            s.setString(16, c.billing_name());
-            s.setString(17, c.billing_phone());
-            s.setString(18, c.billing_country());
-            s.setString(19, c.billing_suburb());
-            s.setString(20, c.billing_city());
-            s.setString(21, c.billing_locality_id());
-            s.setString(22, c.billing_street());
-            s.setString(23, c.billing_postal_code());
-            s.setString(24, c.liable_person());
-            s.setString(25, c.shipping_country());
-            s.setString(26, c.shipping_suburb());
-            s.setString(27, c.shipping_city());
-            s.setString(28, c.shipping_locality_id());
-            s.setString(29, c.shipping_street());
-            s.setString(30, c.shipping_postal_code());
-            s.setString(31, c.shipping_contact());
-            s.setString(32, c.shipping_phone());
-            s.setTimestamp(33, toTimestamp(c.modified()));
-            s.setInt(34, c.id());
-            return s.executeUpdate();
-        }
-    }
-
-    private static int insertLockerDetails(Connection db, LockerDetails ld) throws SQLException {
-        try (var s = db.prepareStatement("INSERT INTO locker_details (locker_id, locker_name, locker_delivery_eligible, courier_external_office_id) VALUES (?,?,?,?) ON CONFLICT(locker_id) DO NOTHING")) {
-            s.setString(1, ld.locker_id());
-            s.setString(2, ld.locker_name());
-            s.setInt(3, ld.locker_delivery_eligible());
-            s.setString(4, ld.courier_external_office_id());
-            return s.executeUpdate();
-        }
-    }
-
-    private static LockerDetails selectLockerDetails(Connection db, String lockerId) throws SQLException {
-        LockerDetails lockerDetails = null;
-        try (var s = db.prepareStatement("SELECT * FROM locker_details WHERE locker_id = ?")) {
-            s.setString(1, lockerId);
-            try (var rs = s.executeQuery()) {
-                if (rs.next()) {
-                    lockerDetails = new LockerDetails(rs.getString("locker_id"), rs.getString("locker_name"), rs.getInt("locker_delivery_eligible"), rs.getString("courier_external_office_id"));
-                }
-            }
-        }
-        return lockerDetails;
-    }
-
-    private static int updateLockerDetails(Connection db, LockerDetails ld) throws SQLException {
-        try (var s = db.prepareStatement("UPDATE locker_details SET locker_name = ?, locker_delivery_eligible = ?, courier_external_office_id = ? WHERE locker_id = ?")) {
-            s.setString(1, ld.locker_name());
-            s.setInt(2, ld.locker_delivery_eligible());
-            s.setString(3, ld.courier_external_office_id());
-            s.setString(4, ld.locker_id());
-            return s.executeUpdate();
-        }
-    }
-
-    /**
-     * Record for reporting the result of inserting an order.
-     *
-     * @param inserted true if a record was created, false otherwise.
-     * @param surrogateId the ID of either the freshly created
-     */
-    record InsertResult(boolean inserted, int surrogateId) {
-    }
-
-    /**
-     * Insert an order if possible.
-     *
-     * @param db the database.
-     * @param or the order to insert.
-     * @param vendorId the vendor to which this order belongs.
-     * @return both the status of the operation and if that is true also the surrogate ID generated by the database.
-     * @throws SQLException if anything goes wrong.
-     */
-    private static InsertResult insertOrder(Connection db, OrderResult or, UUID vendorId) throws SQLException {
+    private static Map<String, List<POInfo>> getOrderDataByProduct(Connection db, String productCode) throws SQLException {
+        var result = new HashMap<String, List<POInfo>>();
         try (var s = db.prepareStatement("""
-                INSERT INTO emag_order (
-                    vendor_id,
-                    id,
-                    status,
-                    is_complete,
-                    type,
-                    payment_mode,
-                    payment_mode_id,
-                    delivery_payment_mode,
-                    delivery_mode,
-                    observation,
-                    details_id,
-                    date,
-                    payment_status,
-                    cashed_co,
-                    cashed_cod,
-                    shipping_tax,
-                    customer_id,
-                    is_storno,
-                    cancellation_reason,
-                    refunded_amount,
-                    refund_status,
-                    maximum_date_for_shipment,
-                    finalization_date,
-                    parent_id,
-                    detailed_payment_method,
-                    proforms,
-                    cancellation_request,
-                    has_editable_products,
-                    late_shipment,
-                    emag_club,
-                    weekend_delivery,
-                    created,
-                    modified,
-                    cancellation_reason_text
-                ) VALUES (?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?,?) ON CONFLICT(id, vendor_id, status) DO NOTHING
-                """, RETURN_GENERATED_KEYS)) {
-            s.setObject(1, vendorId);
-            s.setString(2, or.id());
-            s.setInt(3, or.status());
-            s.setInt(4, or.is_complete());
-            s.setInt(5, or.type());
-            s.setString(6, or.payment_mode());
-            s.setInt(7, or.payment_mode_id());
-            s.setString(8, or.delivery_payment_mode());
-            s.setString(9, or.delivery_mode());
-            s.setString(10, or.observation());
-            s.setString(11, or.details().locker_id());
-            s.setTimestamp(12, toTimestamp(or.date()));
-            s.setInt(13, or.payment_status());
-            s.setBigDecimal(14, or.cashed_co());
-            s.setBigDecimal(15, or.cashed_cod());
-            s.setBigDecimal(16, or.shipping_tax());
-            s.setObject(17, or.customer() != null ? or.customer().id() : null);
-            s.setBoolean(18, or.is_storno());
-            s.setObject(19, or.reason_cancellation() == null ? null : or.reason_cancellation().id());
-            s.setBigDecimal(20, or.refunded_amount());
-            s.setString(21, or.refund_status());
-            s.setTimestamp(22, toTimestamp(or.maximum_date_for_shipment()));
-            s.setTimestamp(23, toTimestamp(or.finalization_date()));
-            s.setString(24, or.parent_id());
-            s.setString(25, or.detailed_payment_method());
-            s.setString(26, String.join("", or.proforms()));
-            s.setString(27, or.cancellation_request());
-            s.setInt(28, or.has_editable_products());
-            s.setObject(29, or.late_shipment());
-            s.setInt(30, or.emag_club());
-            s.setInt(31, or.weekend_delivery());
-            s.setTimestamp(32, toTimestamp(or.created()));
-            s.setTimestamp(33, toTimestamp(or.modified()));
-            s.setString(34, or.reason_cancellation() == null ? null : or.reason_cancellation().name());
-            int insertedRows = s.executeUpdate();
-            if (insertedRows == 0) {
-                var surrogateId = selectSurrogateId(db, or.id(), or.vendor_name(), or.status());
-                return new InsertResult(false, surrogateId);
-            } else if (insertedRows == 1) {
-                try (var generatedKeys = s.getGeneratedKeys()) {
-                    if (generatedKeys.next()) {
-                        int surrogateId = generatedKeys.getInt("surrogate_id");
-                        return new InsertResult(true, surrogateId);
-                    } else {
-                        throw new SQLException("Creating order failed, no surrogate ID obtained.");
-                    }
-                }
-            } else {
-                throw new RuntimeException("Unexpected insertion of more than one row! " + or);
-            }
-        }
-    }
-
-    /**
-     * Find the surrogate ID for a given (orderId, vendorName, status).
-     *
-     * @param db the database.
-     * @param orderId the ID from emag.
-     * @param vendorName the name of the vendor.
-     * @param status the status.
-     * @return The surrogate ID or null.
-     * @throws SQLException if anything goes wrong.
-     */
-    private static Integer selectSurrogateId(Connection db, String orderId, String vendorName, int status) throws SQLException {
-        Integer result = null;
-        try (var s = db.prepareStatement("SELECT surrogate_id FROM emag_order AS o INNER JOIN vendor AS v ON o.vendor_id = v.id WHERE o.id = ? AND v.vendor_name = ? AND o.status = ?")) {
-            s.setString(1, orderId);
-            s.setString(2, vendorName);
-            s.setInt(3, status);
+                SELECT p.name AS productName,
+                pio.quantity AS quantity,
+                pio.initial_qty AS initialQuantity,
+                pio.storno_qty AS stornoQuantity,
+                pio.sale_price AS salePrice,
+                pio.vat AS vat,
+                pio.created, pio.modified,
+                o.status AS orderStatus,
+                o.date AS orderDate,
+                o.modified,
+                o.id AS orderId
+                FROM product_in_order AS pio
+                INNER JOIN product AS p ON p.emag_pnk = pio.part_number_key
+                INNER JOIN emag_order AS o ON pio.emag_order_surrogate_id = o.surrogate_id
+                WHERE p.product_code = ? AND (o.status = 4 OR o.status = 5)
+                ORDER BY o.id, o.status
+                """)) {
+            s.setObject(1, productCode);
             try (var rs = s.executeQuery()) {
-                if (rs.next()) {
-                    result = rs.getInt(1);
+                while (rs.next()) {
+                    var orderId = rs.getString("orderId");
+                    var poInfos = result.getOrDefault(orderId, new ArrayList<>());
+                    var quantity = rs.getInt("quantity");
+                    var initialQuantity = rs.getInt("initialQuantity");
+                    var stornoQuantity = rs.getInt("stornoQuantity");
+                    var salePrice = rs.getBigDecimal("salePrice");
+                    var vat = new BigDecimal(rs.getString("vat"));
+                    var price = vat.add(BigDecimal.ONE).multiply(salePrice).setScale(2, HALF_EVEN);
+                    var poInfo = new POInfo(
+                            orderId,
+                            toLocalDate(rs.getTimestamp("orderDate")),
+                            rs.getInt("orderStatus"),
+                            rs.getString("productName"),
+                            quantity,
+                            initialQuantity,
+                            stornoQuantity,
+                            price
+                    );
+                    poInfos.add(poInfo);
+                    result.put(orderId, poInfos);
                 }
-                if (rs.next()) {
-                    throw new RuntimeException("Found two orders with ID=%s, vendor=%s and status=%d.".formatted(orderId, vendorName, status));
-                }
+
             }
         }
         return result;
-    }
-
-    private static OrderResult selectOrder(Connection db,
-                                           int surrogateId,
-                                           String vendorName,
-                                           Customer customer,
-                                           List<VoucherSplit> shipping_tax_voucher_split,
-                                           List<Product> products,
-                                           List<Attachment> attachments,
-                                           List<Voucher> vouchers,
-                                           List<String> enforcedVendorCourierAccounts,
-                                           List<Flag> flags) throws SQLException {
-        OrderResult order = null;
-        try (var s = db.prepareStatement("SELECT * FROM emag_order WHERE surrogate_id = ?")) {
-            s.setInt(1, surrogateId);
-            try (var rs = s.executeQuery()) {
-                if (rs.next()) {
-                    order = new OrderResult(
-                            vendorName, rs.getString("id"), rs.getInt("status"), rs.getInt("is_complete"), rs.getInt("type"), rs.getString("payment_mode"),
-                            rs.getInt("payment_mode_id"), rs.getString("delivery_payment_mode"), rs.getString("delivery_mode"), rs.getString("observation"),
-                            new LockerDetails(rs.getString("details_id"), null, 0, null), // Assuming LockerDetails has a constructor that takes locker_id
-                            toLocalDateTime(rs.getTimestamp("date")), rs.getInt("payment_status"), rs.getBigDecimal("cashed_co"), rs.getBigDecimal("cashed_cod"),
-                            rs.getBigDecimal("shipping_tax"),
-                            shipping_tax_voucher_split, customer, products, attachments, vouchers,
-                            rs.getBoolean("is_storno"),
-                            rs.getBigDecimal("refunded_amount"),
-                            rs.getString("refund_status"),
-                            toLocalDateTime(rs.getTimestamp("maximum_date_for_shipment")),
-                            toLocalDateTime(rs.getTimestamp("finalization_date")),
-                            rs.getString("parent_id"),
-                            rs.getString("detailed_payment_method"),
-                            Arrays.asList(rs.getString("proforms").split("\n")), // Split the string back into a list
-                            rs.getString("cancellation_request"),
-                            rs.getInt("has_editable_products"),
-                            new CancellationReason(rs.getObject("cancellation_reason", Integer.class),
-                                    rs.getString("cancellation_reason_text")),
-                            rs.getObject("late_shipment", Integer.class), // Assuming late_shipment is a Timestamp
-                            flags, rs.getInt("emag_club"),
-                            rs.getInt("weekend_delivery"),
-                            toLocalDateTime(rs.getTimestamp("created")),
-                            toLocalDateTime(rs.getTimestamp("modified")),
-                            enforcedVendorCourierAccounts
-                    );
-                }
-            }
-        }
-        return order;
-    }
-    private static int insertAWB(Connection db, AWB awb, int emagId) throws SQLException {
-        try (var s = db.prepareStatement("""
-                INSERT INTO awb (
-                reservation_id,
-                emag_id
-                ) VALUES (?, ?) ON CONFLICT(reservation_id) DO NOTHING""")) {
-            s.setInt(1, awb.reservation_id());
-            s.setInt(2, emagId);
-            return s.executeUpdate();
-        }
-    }
-
-    private static int insertRequestHistory(Connection db, RequestHistory requestHistory, int emagId) throws SQLException {
-        try (var s = db.prepareStatement("""
-                INSERT INTO request_history (
-                emag_id,
-                id,
-                req_user,
-                action,
-                action_type,
-                source,
-                date
-                ) VALUES (?, ?, ?, ?, ?, ?, ?) ON CONFLICT(id) DO NOTHING""")) {
-            s.setInt(1, emagId);
-            s.setLong(2,requestHistory.id());
-            s.setString(3,requestHistory.user());
-            s.setString(4,requestHistory.action());
-            s.setString(5,requestHistory.action_type());
-            s.setString(6,requestHistory.source());
-            s.setTimestamp(7, toTimestamp(requestHistory.date()));
-            return s.executeUpdate();
-        }
-    }
-
-    private static int insertStatusHistory(Connection db, StatusHistory statusHistory, int emagId, UUID uuid) throws SQLException {
-        try (var s = db.prepareStatement("INSERT INTO status_history (uuid, code, event_date, emag_id) VALUES (?, ?, ?, ?) ON CONFLICT(uuid) DO NOTHING")) {
-            s.setObject(1, uuid);
-            s.setString(2, statusHistory.code());
-            s.setTimestamp(3, statusHistory.event_date() == null ? null : toTimestamp(statusHistory.event_date()));
-            s.setInt(4, emagId);
-            return s.executeUpdate();
-        }
-    }
-
-    private static int insertStatusRequest(Connection db, StatusRequest statusRequest, UUID statusHistoryUuid) throws SQLException {
-        try (var s = db.prepareStatement("INSERT INTO status_request (amount, created, refund_type, refund_status, rma_id, status_date, status_history_uuid) VALUES (?, ?, ?, ?, ?, ?, ?)")) {
-            s.setBigDecimal(1, statusRequest.amount());
-            s.setTimestamp(2, toTimestamp(statusRequest.created()));
-            s.setString(3, statusRequest.refund_type());
-            s.setString(4, statusRequest.refund_status());
-            s.setString(5, statusRequest.rma_id());
-            s.setTimestamp(6, toTimestamp(statusRequest.status_date()));
-            s.setObject(7, statusHistoryUuid);
-            return s.executeUpdate();
-        }
-    }
-    private static int insertReturnedProduct(Connection db, ReturnedProduct returnedProduct, int emagId) throws SQLException {
-        try (var s = db.prepareStatement("""
-                INSERT INTO emag_returned_products (
-                id,
-                product_emag_id,
-                product_id, quantity,
-                product_name,
-                return_reason,
-                observations,
-                diagnostic,
-                reject_reason,
-                retained_amount,
-                emag_id
-                ) VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?) ON CONFLICT(id) DO NOTHING""")) {
-            s.setInt(1, returnedProduct.id());
-            s.setInt(2, returnedProduct.product_emag_id());
-            s.setInt(3, returnedProduct.product_id());
-            s.setInt(4, returnedProduct.quantity());
-            s.setString(5, returnedProduct.product_name());
-            s.setInt(6, returnedProduct.return_reason());
-            s.setString(7, returnedProduct.observations());
-            s.setString(8, returnedProduct.diagnostic());
-            s.setObject(9, returnedProduct.reject_reason());
-            s.setInt(10, returnedProduct.retained_amount());
-            s.setInt(11, emagId);
-            return s.executeUpdate();
-        }
-    }
-
-    private static int insertRMAResult(Connection db, RMAResult rmaResult) throws SQLException {
-        try (var s = db.prepareStatement("""
-                INSERT INTO rma_result (
-                is_full_fbe,
-                emag_id,
-                return_parent_id,
-                order_id,
-                type,
-                is_club,
-                is_fast,
-                customer_name,
-                customer_company,
-                customer_phone,
-                pickup_country,
-                pickup_suburb,
-                pickup_city,
-                pickup_address,
-                pickup_zipcode,
-                pickup_locality_id,
-                pickup_method,
-                customer_account_iban,
-                customer_account_bank,
-                customer_account_beneficiary,
-                replacement_product_emag_id,
-                replacement_product_id,
-                replacement_product_name,
-                replacement_product_quantity,
-                observations,
-                request_status,
-                return_type,
-                return_reason,
-                date,
-                maximum_finalization_date,
-                first_pickup_date,
-                estimated_product_pickup,
-                estimated_product_reception,
-                return_tax_value,
-                swap,
-                return_address_snapshot,
-                locker_hash,
-                locker_pin,
-                locker_pin_interval_end,
-                return_address_id,
-                country,
-                address_type,
-                request_status_reason
-                ) VALUES (?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?, ?)
-                 ON CONFLICT(emag_id) DO NOTHING""")) {
-            s.setInt(1, rmaResult.is_full_fbe());
-            s.setInt(2, rmaResult.emag_id());
-            s.setObject(3, rmaResult.return_parent_id());
-            s.setString(4, rmaResult.order_id());
-            s.setInt(5, rmaResult.type());
-            s.setInt(6, rmaResult.is_club());
-            s.setInt(7, rmaResult.is_fast());
-            s.setString(8, rmaResult.customer_name());
-            s.setString(9, rmaResult.customer_company());
-            s.setString(10, rmaResult.customer_phone());
-            s.setString(11, rmaResult.pickup_country());
-            s.setString(12, rmaResult.pickup_suburb());
-            s.setString(13, rmaResult.pickup_city());
-            s.setString(14, rmaResult.pickup_address());
-            s.setString(15, rmaResult.pickup_zipcode());
-            s.setInt(16, rmaResult.pickup_locality_id());
-            s.setInt(17, rmaResult.pickup_method());
-            s.setString(18, rmaResult.customer_account_iban());
-            s.setString(19, rmaResult.customer_account_bank());
-            s.setString(20, rmaResult.customer_account_beneficiary());
-            s.setObject(21, rmaResult.replacement_product_emag_id());
-            s.setObject(22, rmaResult.replacement_product_id());
-            s.setString(23, rmaResult.replacement_product_name());
-            s.setObject(24, rmaResult.replacement_product_quantity());
-            s.setString(25, rmaResult.observations());
-            s.setInt(26, rmaResult.request_status());
-            s.setInt(27, rmaResult.return_type());
-            s.setInt(28, rmaResult.return_reason());
-            s.setTimestamp(29, toTimestamp(rmaResult.date()));
-            s.setTimestamp(30, rmaResult.extra_info() == null ? null : toTimestamp(rmaResult.extra_info().maximum_finalization_date()));
-            s.setTimestamp(31, rmaResult.extra_info() == null ? null : toTimestamp(rmaResult.extra_info().first_pickup_date()));
-            s.setTimestamp(32, rmaResult.extra_info() == null ? null : toTimestamp(rmaResult.extra_info().estimated_product_pickup()));
-            s.setTimestamp(33, rmaResult.extra_info() == null ? null : toTimestamp(rmaResult.extra_info().estimated_product_reception()));
-            s.setString(34, rmaResult.return_tax_value());
-            s.setString(35, rmaResult.swap());
-            s.setString(36, rmaResult.return_address_snapshot());
-            s.setString(37, rmaResult.locker() == null ? null : rmaResult.locker().locker_hash());
-            s.setString(38, rmaResult.locker() == null ? null : rmaResult.locker().locker_pin());
-            s.setTimestamp(39, rmaResult.locker() == null ? null : toTimestamp(rmaResult.locker().locker_pin_interval_end()));
-            s.setObject(40, rmaResult.return_address_id());
-            s.setString(41, rmaResult.country());
-            s.setString(42, rmaResult.address_type());
-            s.setObject(43, rmaResult.request_status_reason());
-            return s.executeUpdate();
-        }
     }
 }
